+++ tools/src/tester/types/story.types.ts
@@ -38,6 +38,14 @@ export type Parameter = (string | number | boolean)[] | string | number | boolea
  */
 export type Payload = {} | any[] | string | number | boolean;
 /**
+ * The semver range to execute the chapter against.
+ *
+ *
+ * This interface was referenced by `Story`'s JSON-Schema
+ * via the `definition` "Version".
+ */
+export type Version = string;
+/**
  * This interface was referenced by `Story`'s JSON-Schema
  * via the `definition` "Chapter".
  */
@@ -79,6 +87,7 @@ export interface ChapterRequest {
   };
   request_body?: RequestBody;
   output?: Output;
+  version?: Version;
 }
 /**
  * This interface was referenced by `Story`'s JSON-Schema