+++ tools/tests/tester/integ/TestRunner.test.ts
@@ -11,7 +11,11 @@ import { construct_tester_components, flatten_errors, load_expected_evaluation }
 import { type StoryEvaluation } from 'tester/types/eval.types'
 
 test('stories folder', async () => {
-  const { test_runner } = construct_tester_components('tools/tests/tester/fixtures/specs/excerpt.yaml')
+  const { test_runner, opensearch_http_client } = construct_tester_components('tools/tests/tester/fixtures/specs/excerpt.yaml')
+
+  const info = await opensearch_http_client.wait_until_available()
+  expect(info.version).toBeDefined()
+
   const result = await test_runner.run('tools/tests/tester/fixtures/stories')
 
   expect(result.failed).toBeTruthy()